hide demo-specific logic in chat example
